+++ controller/api/bids.js
@@ -1,0 +1,10 @@
+const bids = require("express").Router()
+
+
+
+
+
+
+
+
+module.exports = bids
