+++ server.js
@@ -5,7 +5,7 @@ const sequelize = require("./config/connection")
 const { Ticket } = require("./models/index")
 const exphbs = require("express-handlebars")
 const hbs = exphbs.create({})
-const routes = require("./controller/api")
+const router = require("./controller/index.js")
 
 const app = express()
 const PORT = process.env.PORT || 3001
@@ -27,20 +27,11 @@ app.use(express.json())
 app.use(express.urlencoded({ extended: false}))
 app.engine("handlebars", hbs.engine)
 app.set("view engine", "handlebars")
+app.use(router)
 
 app.use(express.static("public"))
 
-app.get("/", async (req, res) => {
-    try {
-        const ticketData = await Ticket.findAll()
-        const tickets = ticketData.map(ticket => ticket.get({ plain: true}))
-        res.render("home", { tickets : tickets, user: req.session.username })
-    } catch (error) {
-        res.status(500).json(error)
-    }
-})
-
-sequelize.sync().then(
+sequelize.sync({ force: true }).then(
     app.listen(PORT, () => {
         console.log(`Listening at http://localhost:${PORT}`)
     })
